requirements.txt
autopep8==2.3.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pycodestyle==2.12.0
pytesseract==0.3.10
python-dotenv==1.0.1
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
